add login as a dialog
