+++ packages/lesmis/src/components/ProfileButton.js
@@ -7,6 +7,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { useQueryClient } from 'react-query';
 import { ProfileButton as BaseProfileButton, ProfileButtonItem } from '@tupaia/ui-components';
+import { LoginDialog } from './LoginDialog';
 import { post, useUser } from '../api';
 
 const StyledProfileButton = styled(BaseProfileButton)`
@@ -18,15 +19,17 @@ const StyledProfileButton = styled(BaseProfileButton)`
 `;
 
 export const ProfileButton = () => {
-  const { data: user } = useUser();
+  const { data: user, isSuccess } = useUser();
   const queryClient = useQueryClient();
 
   const handleLogout = async () => {
     await post('logout');
     queryClient.invalidateQueries('user');
+    queryClient.invalidateQueries('entity');
+    queryClient.invalidateQueries('entities');
   };
 
-  return (
+  return user && isSuccess ? (
     <StyledProfileButton
       user={{ ...user, name: `${user.firstName} ${user.lastName}` }}
       MenuOptions={() => (
@@ -35,5 +38,7 @@ export const ProfileButton = () => {
         </ProfileButtonItem>
       )}
     />
+  ) : (
+    <LoginDialog buttonText="Log in" />
   );
 };
